backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-multipart==0.0.17
aiosqlite==0.20.0
openai>=1.0.0
python-dotenv>=1.0.0
